requirements.txt
numpy
pandas
scikit-learn
jupyter
ipykernel
python-dotenv
pyyaml
joblib
fastapi
uvicorn[standard]